refactor: featurization
