+++ ppmat/models/polymer_chemprop/featurization.py
@@ -1,4 +1,3 @@
-from collections import Counter
 from copy import deepcopy
 from dataclasses import dataclass
 from itertools import zip_longest
@@ -10,65 +9,11 @@ import numpy as np
 import paddle
 from rdkit import Chem
 
-# ---------------------------------------------------------------------------
-# RDKit helpers (ported from chemprop/rdkit.py)
-# ---------------------------------------------------------------------------
-
-
-def make_mol(s: str, keep_h: bool, add_h: bool):
-    """
-    Builds an RDKit molecule from a SMILES string.
-
-    :param s: SMILES string.
-    :param keep_h: Boolean whether to keep hydrogens in the input smiles.
-    :param add_h: Boolean whether to add hydrogens.
-    :return: RDKit molecule.
-    """
-    if keep_h:
-        mol = Chem.MolFromSmiles(s, sanitize=False)
-        Chem.SanitizeMol(
-            mol,
-            sanitizeOps=Chem.SanitizeFlags.SANITIZE_ALL
-            ^ Chem.SanitizeFlags.SANITIZE_ADJUSTHS,
-        )
-    else:
-        mol = Chem.MolFromSmiles(s)
-    if add_h:
-        mol = Chem.AddHs(mol)
-    return mol
-
-
-def make_polymer_mol(smiles: str, keep_h: bool, add_h: bool, fragment_weights: list):
-    """
-    Builds an RDKit molecule from a polymer SMILES string with fragment weights.
-
-    :param smiles: SMILES string (fragments separated by '.').
-    :param keep_h: Boolean whether to keep hydrogens in the input smiles.
-    :param add_h: Boolean whether to add hydrogens.
-    :param fragment_weights: List of monomer fractions for each fragment.
-    :return: RDKit molecule.
-    """
-    num_frags = len(smiles.split("."))
-    if len(fragment_weights) != num_frags:
-        raise ValueError(
-            f"number of input monomers/fragments ({num_frags}) does not match number of "
-            f"input number of weights ({len(fragment_weights)})"
-        )
-
-    mols = []
-    for s, w in zip(smiles.split("."), fragment_weights):
-        m = make_mol(s, keep_h, add_h)
-        for a in m.GetAtoms():
-            a.SetDoubleProp("w_frag", float(w))
-        mols.append(m)
-
-    mol = mols.pop(0)
-    while len(mols) > 0:
-        m2 = mols.pop(0)
-        mol = Chem.CombineMols(mol, m2)
-
-    return mol
-
+from ppmat.models.polymer_chemprop.rdkit import make_mol
+from ppmat.models.polymer_chemprop.rdkit import make_polymer_mol
+from ppmat.models.polymer_chemprop.rdkit import parse_polymer_rules
+from ppmat.models.polymer_chemprop.rdkit import remove_wildcard_atoms
+from ppmat.models.polymer_chemprop.rdkit import tag_atoms_in_repeating_unit
 
 # ---------------------------------------------------------------------------
 # Featurization configuration (replaces global PARAMS)
@@ -76,7 +21,7 @@ def make_polymer_mol(smiles: str, keep_h: bool, add_h: bool, fragment_weights: l
 
 
 @dataclass
-class FeaturizationConfig:
+class Featurization_parameters:
     max_atomic_num: int = 100
     explicit_h: bool = False
     add_h: bool = False
@@ -110,10 +55,10 @@ class FeaturizationConfig:
 _DEFAULT_CONFIG = None
 
 
-def _get_default_config() -> FeaturizationConfig:
+def _get_default_config() -> Featurization_parameters:
     global _DEFAULT_CONFIG
     if _DEFAULT_CONFIG is None:
-        _DEFAULT_CONFIG = FeaturizationConfig()
+        _DEFAULT_CONFIG = Featurization_parameters()
     return _DEFAULT_CONFIG
 
 
@@ -123,7 +68,7 @@ def _get_default_config() -> FeaturizationConfig:
 
 
 def get_atom_fdim(
-    config: FeaturizationConfig = None, overwrite_default_atom: bool = False
+    config: Featurization_parameters = None, overwrite_default_atom: bool = False
 ) -> int:
     if config is None:
         config = _get_default_config()
@@ -131,7 +76,7 @@ def get_atom_fdim(
 
 
 def get_bond_fdim(
-    config: FeaturizationConfig = None,
+    config: Featurization_parameters = None,
     atom_messages: bool = False,
     overwrite_default_bond: bool = False,
     overwrite_default_atom: bool = False,
@@ -164,7 +109,7 @@ def onek_encoding_unk(value: int, choices: List[int]) -> List[int]:
 def atom_features(
     atom: Chem.rdchem.Atom,
     functional_groups: List[int] = None,
-    config: FeaturizationConfig = None,
+    config: Featurization_parameters = None,
 ) -> List[Union[bool, int, float]]:
     """
     Builds a feature vector for an atom.
@@ -200,7 +145,7 @@ def atom_features(
 
 
 def bond_features(
-    bond: Chem.rdchem.Bond, config: FeaturizationConfig = None
+    bond: Chem.rdchem.Bond, config: Featurization_parameters = None
 ) -> List[Union[bool, int, float]]:
     """
     Builds a feature vector for a bond.
@@ -225,83 +170,6 @@ def bond_features(
 
 
 # ---------------------------------------------------------------------------
-# Polymer helpers
-# ---------------------------------------------------------------------------
-
-
-def tag_atoms_in_repeating_unit(mol):
-    """
-    Tags atoms that are part of the core units, as well as atoms serving to identify
-    attachment points. In addition, create a map of bond types based on what bonds are
-    connected to R groups in the input.
-    """
-    atoms = [a for a in mol.GetAtoms()]
-    neighbor_map = {}
-    r_bond_types = {}
-
-    for atom in atoms:
-        if "*" in atom.GetSmarts():
-            neighbors = atom.GetNeighbors()
-            assert len(neighbors) == 1
-            neighbor_idx = neighbors[0].GetIdx()
-            r_tag = atom.GetSmarts().strip("[]").replace(":", "")
-            neighbor_map[r_tag] = neighbor_idx
-            atom.SetBoolProp("core", False)
-            bond = mol.GetBondBetweenAtoms(atom.GetIdx(), neighbor_idx)
-            r_bond_types[r_tag] = bond.GetBondType()
-        else:
-            atom.SetBoolProp("core", True)
-
-    for atom in atoms:
-        if atom.GetIdx() in neighbor_map.values():
-            r_tags = [k for k, v in neighbor_map.items() if v == atom.GetIdx()]
-            atom.SetProp("R", "".join(r_tags))
-        else:
-            atom.SetProp("R", "")
-
-    return mol, r_bond_types
-
-
-def remove_wildcard_atoms(rwmol):
-    indices = [a.GetIdx() for a in rwmol.GetAtoms() if "*" in a.GetSmarts()]
-    while len(indices) > 0:
-        rwmol.RemoveAtom(indices[0])
-        indices = [a.GetIdx() for a in rwmol.GetAtoms() if "*" in a.GetSmarts()]
-    Chem.SanitizeMol(rwmol, Chem.SanitizeFlags.SANITIZE_ALL)
-    return rwmol
-
-
-def parse_polymer_rules(rules):
-    polymer_info = []
-    counter = Counter()
-
-    if "~" in rules[-1]:
-        Xn = float(rules[-1].split("~")[1])
-        rules[-1] = rules[-1].split("~")[0]
-    else:
-        Xn = 1.0
-
-    for rule in rules:
-        if rule == "":
-            continue
-        if len(rule.split(":")) != 3:
-            raise ValueError(f'incorrect format for input information "{rule}"')
-        idx1, idx2 = rule.split(":")[0].split("-")
-        w12 = float(rule.split(":")[1])
-        w21 = float(rule.split(":")[2])
-        polymer_info.append((idx1, idx2, w12, w21))
-        counter[idx1] += float(w21)
-        counter[idx2] += float(w12)
-
-    for k, v in counter.items():
-        if np.isclose(v, 1.0) is False:
-            raise ValueError(
-                f"sum of weights of incoming stochastic edges should be 1 -- found {v} for [*:{k}]"
-            )
-    return polymer_info, 1.0 + np.log10(Xn)
-
-
-# ---------------------------------------------------------------------------
 # MolGraph
 # ---------------------------------------------------------------------------
 
@@ -318,7 +186,7 @@ class MolGraph:
         bond_features_extra: np.ndarray = None,
         overwrite_default_atom_features: bool = False,
         overwrite_default_bond_features: bool = False,
-        config: FeaturizationConfig = None,
+        config: Featurization_parameters = None,
     ):
         """
         :param mol: A SMILES or an RDKit molecule.
@@ -326,7 +194,7 @@ class MolGraph:
         :param bond_features_extra: Additional bond features as numpy array.
         :param overwrite_default_atom_features: Whether to overwrite default atom features.
         :param overwrite_default_bond_features: Whether to overwrite default bond features.
-        :param config: A FeaturizationConfig instance (uses default if None).
+        :param config: A Featurization_parameters instance (uses default if None).
         """
         if config is None:
             config = _get_default_config()
@@ -724,7 +592,7 @@ def mol2graph(
     bond_features_batch: List[np.array] = (None,),
     overwrite_default_atom_features: bool = False,
     overwrite_default_bond_features: bool = False,
-    config: FeaturizationConfig = None,
+    config: Featurization_parameters = None,
 ) -> BatchMolGraph:
     """
     Converts a list of SMILES or RDKit molecules to a BatchMolGraph.
@@ -734,7 +602,7 @@ def mol2graph(
     :param bond_features_batch: A list of 2D numpy arrays with additional bond features.
     :param overwrite_default_atom_features: Whether to overwrite default atom descriptors.
     :param overwrite_default_bond_features: Whether to overwrite default bond descriptors.
-    :param config: A FeaturizationConfig instance (uses default if None).
+    :param config: A Featurization_parameters instance (uses default if None).
     :return: A BatchMolGraph containing the combined molecular graph.
     """
     return BatchMolGraph(
